requirements.txt
django==3.2.*
psycopg2-binary==2.9.*
python-dotenv==0.20.*
dj-database-url==0.5.*
environs==9.5.*